Transfer aHasKidsCustomString to aHasKidsAttributes in nested nav
The dspPrimaryNav function in templates/inc/navbar.cfm uses a custom attribute "aHasKidsCustomString", allowing for the dropdown "data-toggle" attribute for bootstrap dropdown menus. In my initial setup, the string was not being applied, even though it is in navbar.cfm (aHasKidsCustomString='role="button" data-toggle="dropdown" data-target="##"')

The result was no dropdown menus. 
NOT SURE IF THIS IS THE RIGHT LOCATION for this change, but in my case it did the trick.