remove min-height on .mura-editable .inactive
When the body content is empty and while logged in as an admin the body area has a min-height. This can be troublesome when developing as the min-height is not an accurate display of your css (causes odd spacing and alignment issues). This line removes the min-height declaration until the content contributor has actually clicked 'quick edit' (removing the inactive class).