Slight tweak to listen to type and subtype
This change should fix the QoQ at the end. Before, the QoQ would come back with records that ignored the type and subtype declarations on the EOFB itself (.setType() and .setSubtype() appeared to be ignored). Here's an example of what reproduced the issue:

        <cfset feed = $.getBean('extendObjectFeedBean')> 
        <cfset feed.setSiteID( $.event('siteId') )> 
        <cfset feed.setType('Custom')>
        <cfset feed.setSubType('Equipment')>
        <cfset feed.setMaxItems(0)> 
        <cfdump var="#feed.getQuery()#">

Before the change produced this output:

http://cl.ly/image/0C0K2Z0f3d0l

Note how "Base" types are being returned despite the subtype being set on the bean.

After the change:

http://cl.ly/image/0542140p0E31