Possible fix for returning duplicate rows
In another issue, we were seeing duplicate rows coming back which corresponded to a row for each attribute defined for a given subtype. I'm not sure adding distinct here is ideal, but it seems to ensure we only get back one row per node of the subtype instead of a row per attribute.