Possible fix for and/or groupings in muraORM beans
I'm not sure if this is a complete solution, but I hit a snag trying to use and/or groupings in MuraORM feed beans. The exact same syntax worked fine on a userFeedBean (and on contentFeedBeans) but not on the muraORM ones. This change got me running again and I thought I'd post it up as a pull request in case others have encountered it. Here's a sample syntax I was trying to get working:

		fb.addParam(
			field = "histStart",
			criteria = now(),
			condition = "lte",
			relationship = "and"
		)
		.addParam(relationship='andOpenGrouping')
			.addParam(
				field = "histEnd",
				criteria = "null",
				condition = "eq"
			)
			.addParam(
				field = "histEnd",
				criteria = now(),
				condition = "gte",
				relationship = "or"
			)
		.addParam(relationship='closeGrouping');

After some digging, the resulting query looked like it was putting out empty parameters or the "and" or "or" keyword in the wrong places. Adding the condition (which I saw in the userFeedBean) and the test for null (which i saw in feedGateway) got me running again on the 6.1 codebase.

Thanks!