Fix for null javax path info
When submitting a mailing list signup form on Lucee 4.5.2, the struct key for "javax.servlet.forward.path_info" may be cast as a null, depending on the server configuration (e.g. - CommandBox embedded server).  

This results in the error "the value from key [javax.servlet.forward.path_info] is NULL"    This ensures that the loop does not error out because of the missing path_info variable.