Remove duplicate code
The argument `siteid` is being defined twice in this call to the initMura function. Once in all lowercase `siteid` and again in camel case `siteID`. I searched within the global.js and global.min.js files and there are no references to the camel case variable. So it does not appear to be needed here.