Issue #2057
This is a little more complicated than it needs to be because the only way ACF provides to add 2 datetimes is by using the + operator. But doing so discards the milliseconds. So I have to use dateAdd() to re-add the milliseconds to our timestamp in order to have millisecond resolution on our timestamp. Also, a 2-digit year was used because a 4-digit year overflowed javascript's 64-bit integer which oddly enough only uses 53 bits.