Feed API sort( "mxpRelevance" ) enhancement
Per the Intuit changes we discussed, if a feed returns multiple articles with the same MXP points/scores, it will pick "at random" from the result set, which often means a really old article is what's getting returned (based on how the database does default sorting).  This change will use "tcontent.releaseDate" as the tie breaker, and if "tcontent.releaseDate" is empty (it's not a required field), it will use tcontent.lastUpdate instead.