Added "keywords" as an option for the dsp_nextN.cfm pagination rendering.
Per conversation w/ Eddie, we’d like to have “keywords” be an option that the dsp_nextN.cfm file honors in the qrystr variable the way it does sortBy, SortDirection, etc.  This was added to the CSAC-EIA project as a hack (I’m not sure who did it), then got used in the MessageBoard plugin also for CSAC on Mura 6.2 and is now in use on SACRS (Mura 7) as well.  If we can make it a legit option in Mura 7 we can:

a) make this a new feature for Mura 7.
b) make CSAC’s code more update-safe, as the hack in their current 6.2 code will be a legit feature in Mura 7 core.
c) fix an issue in the SACRS project (pagination is broken on their Bulletin Board).