Additional changes required for ColdFusion 2018
The `expandPath()` function is returning an unexpected string when the `arguments.dir` parameter contains an absolute path (instead of relative). When the parameter contains a string such as `C:\wwwroot\site\some-directory\` the `expandPath()` function is returning `C:\wwwroot\site\C:\wwwroot\site\some-directory\` which throws an error for the `directoryExists()` function.