Update filebrowser.cfc
Filebrowser had an issue with file paths in the "Site Files" and "Application Root" tabs.  "User Assets" worked just fine.
Simple way to recreate the issue:  Navigate to the file browser > User Assets tab and navigate to a folder that contains images.  The thumbnails do not work and previewing does not function properly.
The file paths being generated were forcing and "/assets" string into the path when it was not needed.