Allow for custom objectconfigurator position options
I could find no way to customize the classes listed in the objectconfigurator custom tag.  AFAIK, this custom tag is not an object that can be extended to a site- or theme-specific location.  Similar options such as 'contentGridStyleMap' (a struct of classes that can be applied specifically to the 'Collection' module) are extended this way, allowing the default setting to be overridden in the contentRenderer.cfm.