Update edittext.cfm to allow for server-side rendering on Text module with 'custom' content source
When extending a 'Text' module, it's currently not possible to change the 'render' parameter from 'client' to 'server' if the content source of the module is set to 'custom'.  This is because edittext.cfm sets the value back to 'client' every time the content source is edited.  There is no reason for this value or the async value to be updated by edittext.cfm as these values are set in the configurator based on the content source selected.  They can be modified in the configurator if desired, but only if they are not hardcoded by edittext.cfm in the core.  This also would allow edittext.cfm to be called by completely custom modules that use a server-side render setting.