Move user credentials to a separate table
In preparation for new changes, move the passwords of users to a separate table. Passswords that are not hashed using bcrypt are dropped and the user needs to create a new password using the 'Forgot password' function. New passwords are now checked to make sure they are not the same as any of the last 6 old password.

Todo:
- fix site bundle support
- fix the expire password after X days mechanism

Incompatible changes:
- unencrypted / MD5 passwords are no longer suported
- legacy sessions are no longer suported