Update configBean.cfc to set a default on mfadayslinkvalid
Adds mfadayslinkvalid default in the configbean. This fixes an error that is thrown when the user does not have mfadayslinkvalid set in settings.ini and a user tries to login via the login link. In that case, this https://github.com/MasaCMS/MasaCMS/blob/main/core/mura/user/userUtility.cfc#L822 will throw because the get method returns an empty string which is not a number. "can't cast empty string to a number value"